* Moving `ImageStatus` to settings file.
