+++ src/panoptes/data/settings.py
@@ -1,3 +1,4 @@
+from enum import IntEnum, auto
 
 from pydantic.networks import AnyHttpUrl
 from pydantic_settings import BaseSettings
@@ -8,3 +9,22 @@ class CloudSettings(BaseSettings):
     img_bucket: str = 'panoptes-images-incoming'
     img_metadata_url: AnyHttpUrl = 'https://us-central1-project-panoptes-01.cloudfunctions.net/get-observation-info'
     observations_url: AnyHttpUrl = 'https://storage.googleapis.com/panoptes-assets/observations.csv'
+
+
+class ImageStatus(IntEnum):
+    """The status of an image."""
+    ERROR = auto()
+    MASKED = auto()
+    UNKNOWN = auto()
+    RECEIVING = auto()
+    RECEIVED = auto()
+    UNSOLVED = auto()
+    PROCESSING = auto()
+    CALIBRATING = auto()
+    CALIBRATED = auto()
+    SOLVING = auto()
+    SOLVED = auto()
+    MATCHING = auto()
+    MATCHED = auto()
+    EXTRACTING = auto()
+    EXTRACTED = auto()
